# app/Http/Requests/CreateInvoiceRequest.php
<?php
/**
 * Created by PhpStorm.
 * User: kirya
 * Date: 13.09.16
 * Time: 16:24
 */

namespace App\Http\Requests;



use App\Demand\ResponseItem;
use App\Services\ResponseItemService;
use Swagger\Annotations as SWG;

/**
 * @SWG\Definition(
 *      definition="CreateInvoiceRequest",
 *      required={"responseItems"},
 *      @SWG\Property(
 *          property="responseItems", type="array", description="id of response_item",
 *          @SWG\Items(type="integer")
 *      )
 * )
 * Class CreateInvoiceRequest
 * @package App\Http\Requests
 */
class CreateInvoiceRequest extends ApiRequest
{

    /**
     * @var ResponseItemService
     */
    private $responseItemService;

    /**
     * @var ResponseItem[]
     */
    private $items;

    public function __construct(array $query = array(), array $request = array(), array $attributes = array(), array $cookies = array(), array $files = array(), array $server = array(), $content = null)
    {
        $this->responseItemService = new ResponseItemService();
        parent::__construct($query, $request, $attributes, $cookies, $files, $server, $content); // TODO: Change the autogenerated stub
    }


    /**
     * Determine if the user is authorized to make this request.
     *
     * @return bool
     */
    public function authorize()
    {
        $items = $this->getResponseItemModels();
        foreach ($items as $item) {
            if (!$this->user()->can('update', $item)) {
                return false;
            }
        }
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array
     */
    public function rules()
    {
        return [
            'responseItems' => 'array|required',
            'responseItems.*' => 'int',
        ];
    }


    /**
     * @return \App\Demand\ResponseItem[]|mixed
     */
    public function getResponseItemModels()
    {
        if (!isset($this->items)) {
            $this->items = $this->responseItemService->findItems($this->responseItems);
        }
        return $this->items;
    }

}