# public/doc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Supp",
        "description": "An API that uses supp service",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/invoices": {
            "post": {
                "tags": [
                    "invoice"
                ],
                "summary": "Create invoice",
                "description": "",
                "operationId": "createInvoices",
                "parameters": [
                    {
                        "name": "Invoice",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/CreateInvoiceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "headers": {
                            "location": {
                                "description": "/invoices/1",
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/invoices/{id}": {
            "put": {
                "tags": [
                    "invoice"
                ],
                "summary": "Update invoice and(or) upload file",
                "description": "",
                "operationId": "updateInvoices",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation"
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "invoice"
                ],
                "summary": "Delete invoice",
                "description": "",
                "operationId": "deleteInvoices",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation"
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/invoices/{id}/files/{name}": {
            "get": {
                "tags": [
                    "invoice"
                ],
                "summary": "Download file for invoice with name",
                "description": "",
                "operationId": "downloadInvoice",
                "produces": [
                    "application/octet-stream",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation",
                        "schema": {
                            "description": "get content of file",
                            "type": "file"
                        }
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/responseItems/{id}": {
            "delete": {
                "tags": [
                    "responseItem"
                ],
                "summary": "Delete responseItem",
                "description": "",
                "operationId": "deleteResponseItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation"
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "responseItem"
                ],
                "summary": "Update ResponseItem",
                "description": "",
                "operationId": "updateResponseItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "type": "integer"
                    },
                    {
                        "name": "ResponseItem",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UpdateResponseItemRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful operation"
                    },
                    "404": {
                        "$ref": "#/responses/NotFoundResponse"
                    },
                    "401": {
                        "$ref": "#/responses/NotAuthResponse"
                    },
                    "default": {
                        "$ref": "#/responses/DefaultErrorResponse"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "ErrorModel": {
            "required": [
                "status_code",
                "message"
            ],
            "properties": {
                "status_code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "CreateInvoiceRequest": {
            "required": [
                "responseItems"
            ],
            "properties": {
                "responseItems": {
                    "description": "id of response_item",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "UpdateResponseItemRequest": {
            "required": [
                "price"
            ],
            "properties": {
                "price": {
                    "type": "number"
                }
            }
        },
...
    },
    "responses": {
        "NotAuthResponse": {
            "description": "Not Auth",
            "schema": {
                "$ref": "#/definitions/ErrorModel"
            }
        },
        "NotFoundResponse": {
            "description": "Not Found",
            "schema": {
                "$ref": "#/definitions/ErrorModel"
            }
        },
        "DefaultErrorResponse": {
            "description": "Default Error",
            "schema": {
                "$ref": "#/definitions/ErrorModel"
            }
        }
    },
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}